Indexing plots of innsight_ggplot2

The S4 class innsight_ggplot2 visualizes the results in the form of a matrix, with the output nodes (and also the input layers) in the columns and the selected data points in the rows. With these basic generic indexing functions, the plots of individual rows and columns can be accessed, modified and the overall plot can be adjusted accordingly.

# S4 method for innsight_ggplot2
[(x, i, j, ..., restyle = TRUE, drop = TRUE)

# S4 method for innsight_ggplot2
[[(x, i, j, ..., restyle = TRUE)

# S4 method for innsight_ggplot2
[(x, i, j, ...) <- value

# S4 method for innsight_ggplot2
[[(x, i, j, ...) <- value

Arguments

x

An instance of the S4 class innsight_ggplot2.

i

The numeric (or missing) index for the rows.

j

The numeric (or missing) index for the columns.

...

other unused arguments

restyle

This logical value determines whether the labels and facet stripes remain as they were in the original plot or are adjusted to the subplot accordingly. However, this argument is only used if the innsight_ggplot2 instance is a multiplot, i.e., x@multiplot is TRUE.

drop

unused argument

value

Another instance of the S4 class innsight_ggplot2 but of shape i x j.

Value

[.innsight_ggplot2: Selects only the plots from the i-th rows and j-th columns and returns them as a new instance of innsight_ggplot2. If restyle = TRUE the facet stripes and axis labels of the original plot are transferred to the subplot, otherwise they are returned as they are.

[[.innsight_ggplot2: Selects only the subplot in row i and column j and returns it as a ggplot2::ggplot object. If restyle = TRUE the facet stripes and axis labels of the original plot are transferred to the subplot, otherwise they are returned as they are.

[<-.innsight_ggplot2: Replaces the plots in the rows i and columns j with those from value and returns the modified instance of innsight_ggplot2.

[[<-.innsight_ggplot2: Replaces the plot from the i-th row and j-th column with the plot from value and returns the modified instance of innsight_ggplot2.

See also