Generic print, plot and show for innsight_plotly

The class innsight_plotly provides the generic visualization functions print, plot and show, which all behave the same in this case. They create a plot of the results using plotly::subplot (see innsight_plotly for details) and return it invisibly.

# S4 method for innsight_plotly
print(x, shareX = TRUE, ...)

# S4 method for innsight_plotly
show(object)

# S4 method for innsight_plotly
plot(x, y, ...)

Arguments

x

An instance of the S4 class innsight_plotly.

shareX

A logical value whether the x-axis should be shared among the subplots.

...

Further arguments passed to plotly::subplot.

object

An instance of the S4 class innsight_plotly.

y

unused argument